Implemented separate navigation graph for GearActivity
